Added concurrency to RPC client
Concurrency is implemented between torrent entries, it syncronises
routines on feeds to improve log readability.
Because of the rate limit from transmission, it was implemented client
rate limiter with 800ms, that might eliminate the benefits of
concurrency. Lock on seen torrent was not implemented, but might be
necessary more tests have to be conducted.